refactor: modularize rendering components
